@@ -22,7 +22,7 @@
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "4.0.0-alpha.60",
-    "@reduxjs/toolkit": "^1.6.2",
+    "@reduxjs/toolkit": "^1.7.1",
     "@rollup/plugin-commonjs": "^17.1.0",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^11.2.1",
@@ -31,9 +31,9 @@
     "@testing-library/user-event": "^12.8.3",
     "@types/enzyme": "^3.10.11",
     "@types/jest": "^27.4.0",
-    "@types/react": "17.0.34",
+    "@types/react": "17.0.38",
     "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.20",
+    "@types/react-redux": "^7.1.22",
     "@types/redux-mock-store": "^1.0.3",
     "axios": "^0.24.0",
     "cross-env": "^7.0.3",
@@ -51,17 +51,17 @@
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.6",
     "react-scripts": "4.0.3",
-    "react-styleguidist": "^11.1.7",
+    "react-styleguidist": "^11.1.8",
     "redux": "^4.1.2",
     "redux-mock-store": "^1.5.4",
     "rollup": "^2.63.0",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-typescript2": "^0.31.1",
-    "ts-jest": "^27.1.1",
+    "ts-jest": "^27.1.2",
     "ts-loader": "^9.2.6",
     "typescript": "^4.5.4",
-    "webpack": "^5.65.0",
+    "webpack": "^5.66.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^3.11.3"
   },
