@@ -9,9 +9,9 @@
     "prettier:format": "prettier -w src/**/*.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.5",
+    "@babel/core": "^7.16.7",
     "@babel/eslint-parser": "^7.16.5",
-    "eslint": "^8.4.1",
+    "eslint": "^8.6.0",
     "prettier": "^2.5.1"
   },
   "packageManager": "yarn@3.1.0"
