@@ -1,9 +1,8 @@
-import * as DOMPurify from 'dompurify';
+import DOMPurify from 'dompurify';
 import ReactHtmlParser, { convertNodeToElement } from 'react-html-parser';
 import * as React from 'react';
 import { ITextResource, IDataSources, ILanguage } from '../types';
-
-const marked = require('marked');
+import marked from 'marked';
 
 DOMPurify.addHook('afterSanitizeAttributes', (node) => {
   if (node.tagName === 'A') {
@@ -56,7 +55,6 @@ export const getParsedLanguageFromText = (text: string, allowedTags?: string[], 
   return parsedText;
 };
 
-// eslint-disable-next-line consistent-return
 const removeStyling = (node: any): React.ReactElement | void | null => {
   // all this does is remove the default styling of the <p> element, which is causing styling issues
   if (node.name === 'p') {
@@ -102,7 +100,6 @@ export function replaceTextResourceParams(
 
       for (let i = 0; i <= repeatingGroupCount; ++i) {
         replaceValues = [];
-        // eslint-disable-next-line no-loop-func
         resource.variables.forEach((variable) => {
           if (variable.dataSource.startsWith('dataModel')) {
             if (variable.key.indexOf('[{0}]') > -1) {
@@ -116,7 +113,6 @@ export function replaceTextResourceParams(
         const newValue = replaceParameters(resource.unparsedValue, replaceValues);
 
         if (resource.repeating && resource.id.endsWith(`-${i}`)) {
-          // eslint-disable-next-line no-param-reassign
           resource.value = newValue;
         } else if (!resource.repeating && textResources.findIndex((r) => r.id === `${resource.id}-${i}`) === -1) {
           const newId = `${resource.id}-${i}`;
@@ -138,7 +134,6 @@ export function replaceTextResourceParams(
 
       const newValue = replaceParameters(resource.unparsedValue, replaceValues);
       if (resource.value !== newValue) {
-        // eslint-disable-next-line no-param-reassign
         resource.value = newValue;
       }
     }
