@@ -38,7 +38,7 @@ const useStyles = makeStyles({
   button: {
     color: theme.altinnPalette.primary.black,
     background: theme.altinnPalette.primary.blue,
-    textTransform: 'none' as const,
+    textTransform: 'none',
     fontWeight: 400,
     height: 36,
     borderRadius: '0',
@@ -90,16 +90,22 @@ const Confirm = (props: IConfirmProps) => {
   const [instanceMetaObject, setInstanceMetaObject] = React.useState({});
   const [isSubmitting, setIsSubmitting] = React.useState<boolean>(false);
 
-  const applicationMetadata: IApplicationMetadata = useAppSelector(state => state.applicationMetadata.applicationMetadata);
-  const instance = useAppSelector(state => state.instanceData.instance);
-  const language = useAppSelector(state => state.language.language);
-  const parties = useAppSelector(state => state.party.parties);
-  const validations = useAppSelector(state => state.formValidations.validations);
+  const applicationMetadata: IApplicationMetadata = useAppSelector(
+    (state) => state.applicationMetadata.applicationMetadata,
+  );
+  const instance = useAppSelector((state) => state.instanceData.instance);
+  const language = useAppSelector((state) => state.language.language);
+  const parties = useAppSelector((state) => state.party.parties);
+  const validations = useAppSelector(
+    (state) => state.formValidations.validations,
+  );
 
   const routeParams: any = props.match.params;
 
   const { instanceId } = window as Window as IAltinnWindow;
-  const textResources = useAppSelector(state => state.textResources.resources);
+  const textResources = useAppSelector(
+    (state) => state.textResources.resources,
+  );
 
   const isLoading = (): boolean =>
     !attachments ||
@@ -115,7 +121,7 @@ const Confirm = (props: IConfirmProps) => {
       routeParams.partyId,
       routeParams.instanceGuid,
     );
-  // eslint-disable-next-line react-hooks/exhaustive-deps
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   React.useEffect(() => {
@@ -130,8 +136,7 @@ const Confirm = (props: IConfirmProps) => {
       });
       setInstanceMetaObject(obj);
     }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [parties, instance, lastChangedDateTime, applicationMetadata]);
+  }, [parties, instance, lastChangedDateTime, applicationMetadata, language]);
 
   React.useEffect(() => {
     if (instance && instance.data && applicationMetadata) {
