@@ -1,5 +1,7 @@
-/* eslint-disable import/no-named-as-default */
-import { AltinnContentIconFormData, AltinnContentLoader } from 'altinn-shared/components';
+import {
+  AltinnContentIconFormData,
+  AltinnContentLoader,
+} from 'altinn-shared/components';
 import { AxiosError } from 'axios';
 import * as React from 'react';
 import { Redirect } from 'react-router-dom';
@@ -9,9 +11,17 @@ import { ShowTypes } from 'src/shared/resources/applicationMetadata';
 import { startInitialStatelessQueue } from 'src/shared/resources/queue/queueSlice';
 import { ISimpleInstance, PresentationType, ProcessTaskType } from 'src/types';
 import { isStatelessApp } from 'src/utils/appMetadata';
-import { checkIfAxiosError, get, HttpStatusCodes, post } from 'src/utils/networking';
-import { getActiveInstancesUrl, getPartyValidationUrl } from 'src/utils/urlHelper';
-import Form from '../form/containers/Form';
+import {
+  checkIfAxiosError,
+  get,
+  HttpStatusCodes,
+  post,
+} from 'src/utils/networking';
+import {
+  getActiveInstancesUrl,
+  getPartyValidationUrl,
+} from 'src/utils/urlHelper';
+import { Form } from '../form/containers/Form';
 import { updateValidations } from '../form/validation/validationSlice';
 import Instantiate from '../instantiate/containers';
 import InstanceSelection from '../instantiate/containers/InstanceSelection';
@@ -19,37 +29,17 @@ import MissingRolesError from '../instantiate/containers/MissingRolesError';
 import NoValidPartiesError from '../instantiate/containers/NoValidPartiesError';
 
 export default function Entrypoint() {
-  const applicationMetadata = useAppSelector(state => state.applicationMetadata.applicationMetadata);
-  const selectedParty = useAppSelector(state => state.party.selectedParty);
+  const applicationMetadata = useAppSelector(
+    (state) => state.applicationMetadata.applicationMetadata,
+  );
+  const selectedParty = useAppSelector((state) => state.party.selectedParty);
   const [action, setAction] = React.useState<ShowTypes>(null);
   const [partyValidation, setPartyValidation] = React.useState(null);
-  const [activeInstances, setActiveInstances] = React.useState<ISimpleInstance[]>(null);
-  const statelessLoading = useAppSelector(state => state.isLoading.stateless);
-  const formDataError = useAppSelector(state => state.formData.error);
+  const [activeInstances, setActiveInstances] =
+    React.useState<ISimpleInstance[]>(null);
+  const statelessLoading = useAppSelector((state) => state.isLoading.stateless);
+  const formDataError = useAppSelector((state) => state.formData.error);
   const dispatch = useAppDispatch();
-
-  const validatatePartySelection = async () => {
-    if (!selectedParty) {
-      return;
-    }
-    try {
-      const { data } = await post(getPartyValidationUrl(selectedParty.partyId));
-      setPartyValidation(data);
-    } catch (err) {
-      console.error(err);
-      throw new Error('Server did not respond with party validation');
-    }
-  };
-
-  const fetchExistingInstances = async () => {
-    try {
-      const instances = await get(getActiveInstancesUrl(selectedParty.partyId));
-      setActiveInstances(instances || []);
-    } catch (err) {
-      console.error(err);
-      throw new Error('Server did not return active instances');
-    }
-  };
 
   const handleNewInstance = () => {
     setAction('new-instance');
@@ -57,23 +47,47 @@ export default function Entrypoint() {
 
   React.useEffect(() => {
     if (action === 'select-instance' && partyValidation?.valid) {
+      const fetchExistingInstances = async () => {
+        try {
+          const instances = await get(
+            getActiveInstancesUrl(selectedParty.partyId),
+          );
+          setActiveInstances(instances || []);
+        } catch (err) {
+          console.error(err);
+          throw new Error('Server did not return active instances');
+        }
+      };
+
       fetchExistingInstances();
     }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [action, partyValidation]);
+  }, [action, partyValidation, selectedParty]);
 
   React.useEffect(() => {
     if (selectedParty) {
+      const validatatePartySelection = async () => {
+        if (!selectedParty) {
+          return;
+        }
+        try {
+          const { data } = await post(
+            getPartyValidationUrl(selectedParty.partyId),
+          );
+          setPartyValidation(data);
+        } catch (err) {
+          console.error(err);
+          throw new Error('Server did not respond with party validation');
+        }
+      };
+
       validatatePartySelection();
     }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [selectedParty]);
 
   React.useEffect(() => {
     // If user comes back to entrypoint from an active instance we need to clear validation messages
     dispatch(updateValidations({ validations: {} }));
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
+  }, [dispatch]);
 
   React.useEffect(() => {
     if (applicationMetadata) {
@@ -88,22 +102,16 @@ export default function Entrypoint() {
 
   if (partyValidation?.valid === false) {
     if (partyValidation.validParties?.length === 0) {
-      return (
-        <NoValidPartiesError />
-      );
+      return <NoValidPartiesError />;
     }
-    return (
-      <Redirect to={`/partyselection/${HttpStatusCodes.Forbidden}`} />
-    );
+    return <Redirect to={`/partyselection/${HttpStatusCodes.Forbidden}`} />;
   }
 
   // error trying to fetch data, if missing rights we display relevant page
   if (checkIfAxiosError(formDataError)) {
     const axiosError = formDataError as AxiosError;
     if (axiosError.response.status === HttpStatusCodes.Forbidden) {
-      return (
-        <MissingRolesError />
-      );
+      return <MissingRolesError />;
     }
   }
 
@@ -112,7 +120,11 @@ export default function Entrypoint() {
     return <Instantiate />;
   }
 
-  if (action === 'select-instance' && partyValidation?.valid && activeInstances !== null) {
+  if (
+    action === 'select-instance' &&
+    partyValidation?.valid &&
+    activeInstances !== null
+  ) {
     if (activeInstances.length === 0) {
       // no existing instances exist, we start instantiation
       return <Instantiate />;
@@ -151,10 +163,7 @@ export default function Entrypoint() {
   }
 
   return (
-    <Presentation
-      header=''
-      type={ProcessTaskType.Unknown}
-    >
+    <Presentation header='' type={ProcessTaskType.Unknown}>
       <AltinnContentLoader width='100%' height='400'>
         <AltinnContentIconFormData />
       </AltinnContentLoader>
