@@ -1,5 +1,5 @@
+import * as React from 'react';
 import { createTheme, createStyles, FormGroup, WithStyles, withStyles } from '@material-ui/core';
-import React = require('react');
 import altinnTheme from '../theme/altinnStudioTheme';
 
 export interface IAltinnCheckBoxGroupProvidedProps extends WithStyles<typeof styles> {
@@ -11,8 +11,6 @@ export interface IAltinnCheckBoxGroupProvidedProps extends WithStyles<typeof sty
   classes: any;
 }
 
-export interface IAltinnCheckBoxGroupState {
-}
 const theme = createTheme(altinnTheme);
 
 const styles = () => createStyles({
