@@ -1,5 +1,4 @@
-import { Button, createTheme } from '@material-ui/core';
-import { withStyles } from '@material-ui/core/styles';
+import { Button, createTheme, makeStyles } from '@material-ui/core';
 import classNames from 'classnames';
 import * as React from 'react';
 import altinnTheme from '../theme/altinnStudioTheme';
@@ -13,27 +12,24 @@ export interface IAltinnIconButtonComponentProvidedProps {
   id?: any;
 }
 
-export interface IAltinnIconButtonComponentState {
-}
-
 const theme = createTheme(altinnTheme);
 
-const styles = {
+const useStyles = makeStyles({
   dottedBtn: {
-    'minHeight': '60px',
-    'fontWeight': 700,
-    'width': '100%',
-    'color': '#000',
-    'textAlign': 'left' as 'left',
-    'verticalAlign': 'middle',
-    'backgroundColor': 'transparent',
-    'border': '1px dotted ' + theme.altinnPalette.primary.blueDark,
-    'boxShadow': 'none',
-    'borderRadius': '0px',
-    'textTransform': 'none' as 'none',
-    'maxWidth': '170px',
-    'justifyContent': 'right',
-    'fontSize': '16px',
+    minHeight: '60px',
+    fontWeight: 700,
+    width: '100%',
+    color: '#000',
+    textAlign: 'left',
+    verticalAlign: 'middle',
+    backgroundColor: 'transparent',
+    border: '1px dotted ' + theme.altinnPalette.primary.blueDark,
+    boxShadow: 'none',
+    borderRadius: '0px',
+    textTransform: 'none',
+    maxWidth: '170px',
+    justifyContent: 'right',
+    fontSize: '16px',
     '&:hover': {
       backgroundColor: 'transparent !Important',
     },
@@ -46,19 +42,27 @@ const styles = {
     fontSize: '54px',
     paddingRight: '6px',
   },
+});
+
+export const AltinnIconButton = ({
+  id,
+  onclickFunction,
+  btnText,
+  iconClass,
+}: IAltinnIconButtonComponentProvidedProps) => {
+  const classes = useStyles();
+
+  return (
+    <Button
+      id={id}
+      variant='contained'
+      className={classes.dottedBtn}
+      onClick={onclickFunction}
+    >
+      <i className={classNames(iconClass, classes.dottedBtnIcon)} />
+      {btnText}
+    </Button>
+  );
 };
 
-
-export class AltinnIconButton extends React.Component<IAltinnIconButtonComponentProvidedProps, IAltinnIconButtonComponentState> {
-  public render() {
-    const { classes } = this.props;
-    return (
-      <Button id={this.props.id} variant='contained' className={classes.dottedBtn} onClick={this.props.onclickFunction}>
-        <i className={classNames(this.props.iconClass, classes.dottedBtnIcon)} />
-        {this.props.btnText}
-      </Button>
-    );
-  }
-}
-
-export default withStyles(styles)(AltinnIconButton);
+export default AltinnIconButton;
