@@ -3,7 +3,7 @@ import * as React from 'react';
 export function AltinnContentIcon() {
   return (
     <>
-      <rect x='0' y='20' rx='0' ry='0' width='100' height='5' />
+      <rect data-testid="AltinnContentIcon" x='0' y='20' rx='0' ry='0' width='100' height='5' />
       <rect x='0' y='30' rx='0' ry='0' width='400' height='5' />
       <rect x='0' y='40' rx='0' ry='0' width='400' height='25' />
 
