@@ -11,7 +11,7 @@ import classNames from 'classnames';
 import * as React from 'react';
 import { IAttachment } from '../../types';
 import { AltinnIcon } from '../AltinnIcon';
-import AltinnAttachment from '../atoms/AltinnAttachment';
+import AltinnAttachmentComponent from '../atoms/AltinnAttachment';
 
 const styles = createStyles({
   listItemTextPadding: {
@@ -77,7 +77,7 @@ export function AltinnCollapsibleAttachments(props: IAltinnCollapsibleAttachment
             />
           </ListItem>
           <Collapse in={open} timeout='auto' unmountOnExit={true}>
-            <AltinnAttachment
+            <AltinnAttachmentComponent
               attachments={props.attachments}
               nested={true}
               listDisableVerticalPadding={true}
@@ -89,7 +89,7 @@ export function AltinnCollapsibleAttachments(props: IAltinnCollapsibleAttachment
           <Typography style={{ fontSize: 18, fontWeight: 600 }}>
             {`${props.title} ${attachmentCount}`}
           </Typography>
-          <AltinnAttachment
+          <AltinnAttachmentComponent
             attachments={props.attachments}
             nested={false}
             listDisableVerticalPadding={false}
