@@ -14,7 +14,7 @@ export const altinnStudioDocsUrl = 'https://altinn.github.io/docs/altinn-studio/
 export const altinnImgLogoHeaderUrl = 'https://altinncdn.no/img/altinn_logo_header.png';
 export const dataModelUploadPageUrl = `${origin}/designer/${org}/${app}#/datamodel`;
 export const dataModelXsdUrl = `${origin}/designer/${org}/${app}/Model/GetXsd`;
-export const orgsListUrl: string = 'https://altinncdn.no/orgs/altinn-orgs.json';
+export const orgsListUrl = 'https://altinncdn.no/orgs/altinn-orgs.json';
 export const repositoryGitUrl = `${origin}/repos/${org}/${app}.git`;
 export const repositoryUrl = `${origin}/repos/${org}/${app}`;
 export const baseHostnameAltinnProd = 'altinn.no';
