@@ -1,26 +1,27 @@
 using System;
+
 using Microsoft.Extensions.Options;
 
 namespace Designer.Tests.Utils
 {
-public class TestOptionsMonitor<T> : IOptionsMonitor<T>
-    where T : class, new()
-{
-    public TestOptionsMonitor(T currentValue)
+    public class TestOptionsMonitor<T> : IOptionsMonitor<T>
+        where T : class, new()
     {
-        CurrentValue = currentValue;
-    }
+        public TestOptionsMonitor(T currentValue)
+        {
+            CurrentValue = currentValue;
+        }
 
-    public T Get(string name)
-    {
-        return CurrentValue;
-    }
+        public T Get(string name)
+        {
+            return CurrentValue;
+        }
 
-    public IDisposable OnChange(Action<T, string> listener)
-    {
-        throw new NotImplementedException();
-    }
+        public IDisposable OnChange(Action<T, string> listener)
+        {
+            throw new NotImplementedException();
+        }
 
-    public T CurrentValue { get; }
-}
+        public T CurrentValue { get; }
+    }
 }
