@@ -4,20 +4,15 @@ import { Provider } from 'react-redux';
 import { HashRouter } from 'react-router-dom';
 import App from './App';
 import { run } from './sagas';
-import { store } from './store';
+import { setupStore } from './store';
 
-/**
- * This is the Script that starts the React application
- */
+const store = setupStore();
 
 /**
  * Setup all Sagas to listen to the defined events
  */
 run();
 
-/**
- *
- */
 render(
   <Provider store={store}>
     <HashRouter>
