@@ -1,35 +1,59 @@
-import { combineReducers, Reducer, ReducersMapObject } from 'redux';
-import { dataModellingReducer, IDataModellingState } from 'app-shared/features/dataModelling/sagas';
-import { dataModelsMetadataReducer, IDataModelsMetadataState } from 'app-shared/features/dataModelling/sagas/metadata';
-import handleServiceInformationReducer, { IHandleServiceInformationState } from '../features/administration/handleServiceInformationSlice';
-import handleMergeConflictReducer, { IHandleMergeConflictState } from '../features/handleMergeConflict/handleMergeConflictSlice';
-import appClusterReducer, { IAppClusterState } from '../sharedResources/appCluster/appClusterSlice';
-import appDeploymentReducer, { IAppDeploymentState } from '../sharedResources/appDeployment/appDeploymentSlice';
-import applicationMetadataReducer, { IApplicationMetadataState } from '../sharedResources/applicationMetadata/applicationMetadataSlice';
-import appReleaseReducer, { IAppReleaseState } from '../sharedResources/appRelease/appReleaseSlice';
-import configurationReducer, { IConfigurationState } from '../sharedResources/configuration/configurationSlice';
-import languageReducer, { IFetchedLanguageState } from '../utils/fetchLanguage/languageSlice';
-import repoStatusReducer, { IRepoStatusState } from '../sharedResources/repoStatus/repoStatusSlice';
+import { Reducer, ReducersMapObject } from 'redux';
+import {
+  dataModellingReducer,
+  IDataModellingState,
+} from 'app-shared/features/dataModelling/sagas';
+import {
+  dataModelsMetadataReducer,
+  IDataModelsMetadataState,
+} from 'app-shared/features/dataModelling/sagas/metadata';
+import handleServiceInformationReducer, {
+  IHandleServiceInformationState,
+} from '../features/administration/handleServiceInformationSlice';
+import handleMergeConflictReducer, {
+  IHandleMergeConflictState,
+} from '../features/handleMergeConflict/handleMergeConflictSlice';
+import appClusterReducer, {
+  IAppClusterState,
+} from '../sharedResources/appCluster/appClusterSlice';
+import appDeploymentReducer, {
+  IAppDeploymentState,
+} from '../sharedResources/appDeployment/appDeploymentSlice';
+import applicationMetadataReducer, {
+  IApplicationMetadataState,
+} from '../sharedResources/applicationMetadata/applicationMetadataSlice';
+import appReleaseReducer, {
+  IAppReleaseState,
+} from '../sharedResources/appRelease/appReleaseSlice';
+import configurationReducer, {
+  IConfigurationState,
+} from '../sharedResources/configuration/configurationSlice';
+import languageReducer, {
+  IFetchedLanguageState,
+} from '../utils/fetchLanguage/languageSlice';
+import repoStatusReducer, {
+  IRepoStatusState,
+} from '../sharedResources/repoStatus/repoStatusSlice';
 import userReducer, { IUserState } from '../sharedResources/user/userSlice';
 
 export interface IServiceDevelopmentReducers
   extends IServiceDevelopmentNameSpace<
-  Reducer<IFetchedLanguageState>,
-  Reducer<IHandleMergeConflictState>,
-  Reducer<IHandleServiceInformationState>,
-  Reducer<IApplicationMetadataState>,
-  Reducer<IAppClusterState>,
-  Reducer<IRepoStatusState>,
-  Reducer<IAppReleaseState>,
-  Reducer<IAppDeploymentState>,
-  Reducer<IConfigurationState>,
-  Reducer<IDataModellingState>,
-  Reducer<IDataModelsMetadataState>,
-  Reducer<IUserState>
-  >,
-  ReducersMapObject { }
+      Reducer<IFetchedLanguageState>,
+      Reducer<IHandleMergeConflictState>,
+      Reducer<IHandleServiceInformationState>,
+      Reducer<IApplicationMetadataState>,
+      Reducer<IAppClusterState>,
+      Reducer<IRepoStatusState>,
+      Reducer<IAppReleaseState>,
+      Reducer<IAppDeploymentState>,
+      Reducer<IConfigurationState>,
+      Reducer<IDataModellingState>,
+      Reducer<IDataModelsMetadataState>,
+      Reducer<IUserState>
+    >,
+    ReducersMapObject {}
 
-const reducers = {
+export const rootReducer = {
   languageState: languageReducer,
   handleMergeConflict: handleMergeConflictReducer,
   serviceInformation: handleServiceInformationReducer,
@@ -43,5 +67,3 @@ const reducers = {
   dataModelsMetadataState: dataModelsMetadataReducer,
   userState: userReducer,
 };
-
-export default combineReducers(reducers);
