@@ -1,17 +1,27 @@
-import { createStore, Store } from 'redux';
-import createSagaMiddleware, { SagaMiddleware } from 'redux-saga';
-import setupMiddlewares from 'app-shared/utils/middleware/setupMiddlewares';
-import reducers from '../reducers';
+import { configureStore, combineReducers } from '@reduxjs/toolkit';
+import { setupListeners } from '@reduxjs/toolkit/query';
 
-export const sagaMiddleware: SagaMiddleware<any> = createSagaMiddleware();
-export const store: Store<IServiceDevelopmentState> = configureStore();
+import { rootReducer } from '../reducers';
+import { sagaMiddleware } from '../sagas';
+import type { PreloadedState } from '@reduxjs/toolkit';
 
-function configureStore(): Store<IServiceDevelopmentState> {
-  const enhancer = setupMiddlewares([sagaMiddleware]);
+export const middlewares = [sagaMiddleware];
 
-  return createStore(
-    reducers,
-    undefined,
-    enhancer,
-  );
-}
+const reducer = combineReducers(rootReducer);
+
+export const setupStore = (preloadedState?: PreloadedState<RootState>) => {
+  const store = configureStore({
+    reducer,
+    devTools: process.env.NODE_ENV !== 'production',
+    middleware: (getDefaultMiddleware: () => any[]) =>
+      getDefaultMiddleware().concat(middlewares),
+    preloadedState,
+  });
+  setupListeners(store.dispatch);
+
+  return store;
+};
+
+export type RootState = ReturnType<typeof reducer>;
+export type AppStore = ReturnType<typeof setupStore>;
+export type AppDispatch = AppStore['dispatch'];
