@@ -1,109 +1,66 @@
 const HtmlWebPackPlugin = require('html-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const MonacoPlugin = require('monaco-editor-webpack-plugin');
 const ForkTsCheckerNotifierWebpackPlugin = require('fork-ts-checker-notifier-webpack-plugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
-const path = require('path');
+
+const commonConfig = require('./webpack.common');
 
 module.exports = {
+  ...commonConfig,
   mode: 'development',
   devtool: 'inline-source-map',
-  entry: [ 
-    "core-js/modules/es.object.assign",
-    "core-js/modules/es.array.find-index",
-    "core-js/modules/es.array.find",
-    "./index.tsx"
-  ],
-  output: {
-    path: path.resolve(__dirname, '../dist/app-development'),
-    filename: "app-development.js"
-  },
-  resolve: {
-    extensions: [".ts", ".tsx", ".js", ".jsx", ".css", ".scss"],
-    alias: {
-      // SHARED
-      "app-shared": path.resolve(__dirname, "../shared/"),
-      "@altinn/schema-editor": path.resolve(__dirname, "../packages/schema-editor/src/"),
-      //"ux-editor": path.resolve(__dirname, "../ux-editor/")
-      "utils": path.resolve(__dirname, "utils/"),
-    }
-  },
   performance: {
     hints: 'warning',
   },
   module: {
-    rules: [{
-      test: /\.jsx?/,
-      exclude: /node_modules/,
-      use: {
-        loader: "babel-loader",
-      }
-    },
-    {
-      test: /\.html$/,
-      use: [{
-        loader: "html-loader",
-        options: {
-          minimize: true
-        }
-      }]
-    },
-    {
-      test: /\.scss$/,
-      use: [
-        "style-loader",
-        "css-loader"
-      ]
-    },
-    {
-      test: /\.svg$/,
-      use: {
-        loader: "svg-inline-loader",
-      }
-    },
-    {
-      test: /\.css$/,
-      use: [{
-        loader: MiniCssExtractPlugin.loader,
+    rules: [
+      ...commonConfig.module.rules,
+      {
+        test: /\.html$/,
+        use: [
+          {
+            loader: 'html-loader',
+            options: {
+              minimize: true,
+            },
+          },
+        ],
       },
       {
-        loader: "css-loader",
-        options: {
-          url: false
-        }
+        test: /\.svg$/,
+        use: {
+          loader: 'svg-inline-loader',
+        },
       },
-      ]
-    },
-    {
-      test: /\.tsx?/,
-      use: [
-        { loader: "ts-loader", options: { transpileOnly: true } }
-      ]
-    },
-    {
-      enforce: "pre",
-      test: /\.js$/,
-      loader: "source-map-loader",
-    }
+      {
+        enforce: 'pre',
+        test: /\.js$/,
+        loader: 'source-map-loader',
+      },
+      {
+        test: /\.tsx?/,
+        use: [
+          {
+            loader: 'ts-loader',
+            options: { transpileOnly: true },
+          },
+        ],
+      },
     ],
   },
   plugins: [
+    ...commonConfig.plugins,
     new ForkTsCheckerWebpackPlugin(),
-    new ForkTsCheckerNotifierWebpackPlugin({ title: 'TypeScript', excludeWarnings: false }),
+    new ForkTsCheckerNotifierWebpackPlugin({
+      title: 'TypeScript',
+      excludeWarnings: false,
+    }),
     new HtmlWebPackPlugin({
       template: './public/index.html',
-      filename: 'index.html'
-    }),
-    new MiniCssExtractPlugin({
-      filename: "app-development.css",
-    }),
-    new MonacoPlugin({
-      output: path.join('../app-development', 'js', 'react'),
-      languages: ['typescript', 'javascript', 'csharp']
+      filename: 'index.html',
     }),
   ],
   devServer: {
     historyApiFallback: true,
     disableHostCheck: true,
-  }
-}
+  },
+};
