@@ -1,32 +1,13 @@
 const HtmlWebPackPlugin = require('html-webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const TerserPlugin = require('terser-webpack-plugin');
 const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
-const MonacoPlugin = require('monaco-editor-webpack-plugin');
-const path = require('path');
+
+const commonConfig = require('./webpack.common');
 
 module.exports = {
+  ...commonConfig,
   mode: 'production',
   devtool: false,
-  entry: [
-    "core-js/modules/es.object.assign",
-    "core-js/modules/es.array.find-index",
-    "core-js/modules/es.array.find",
-    "./index.tsx"
-  ],
-  output: {
-    path: path.resolve(__dirname, '../dist/app-development'),
-    filename: "app-development.js",
-  },
-  resolve: {
-    extensions: [".ts", ".tsx", ".js", ".jsx", ".css", ".scss"],
-    alias: {
-      "app-shared": path.resolve(__dirname, "../shared/"),
-      "@altinn/schema-editor": path.resolve(__dirname, "../packages/schema-editor/src/"),
-      //"ux-editor": path.resolve(__dirname, "../ux-editor/")
-      "utils": path.resolve(__dirname, "utils/"),
-    }
-  },
   performance: {
     hints: false,
   },
@@ -39,42 +20,20 @@ module.exports = {
     ],
   },
   module: {
-    rules: [{
-      test: /\.jsx?/,
-      exclude: /node_modules/,
-      use: {
-        loader: "babel-loader",
-      }
-    },
-    {
-      test: /\.scss$/,
-      use: [
-        "style-loader",
-        "css-loader"
-      ]
-    },
-    {
-      test: /\.css$/,
-      use: [{
-        loader: MiniCssExtractPlugin.loader,
-      },
+    rules: [
+      ...commonConfig.module.rules,
       {
-        loader: "css-loader",
-        options: {
-          url: false
-        }
-      }
-      ]
-    },
-    {
-      test: /\.tsx?/,
-      use: [
-        { loader: "ts-loader", options: { transpileOnly: true } }
-      ]
-    }
+        test: /\.tsx?/,
+        use: [
+          {
+            loader: 'ts-loader',
+          },
+        ],
+      },
     ],
   },
   plugins: [
+    ...commonConfig.plugins,
     new ForkTsCheckerWebpackPlugin({
       async: false,
     }),
@@ -94,12 +53,5 @@ module.exports = {
         minifyURLs: true,
       },
     }),
-    new MiniCssExtractPlugin({
-      filename: "app-development.css",
-    }),
-    new MonacoPlugin({
-      output: path.join('../app-development', 'js', 'react'),
-      languages: ['typescript', 'javascript', 'csharp'],
-    }),
   ],
-}
+};
