@@ -1,4 +1,5 @@
-import { validateRepoName } from 'common/utils';
+import { SelectedContextType } from 'app-shared/navigation/main-header/Header';
+import { validateRepoName, userHasAccessToSelectedContext } from 'common/utils';
 
 describe('validateRepoName', () => {
   it('should return true when repo name does not contain invalid characters', () => {
@@ -35,5 +36,63 @@ describe('validateRepoName', () => {
 
   it('should return false when length is more than 30 characters', () => {
     expect(validateRepoName('a234567890123456789012345678901')).toBe(false);
+  });
+});
+
+describe('userHasAccessToSelectedContext', () => {
+  it('should return true when context is self', () => {
+    const result = userHasAccessToSelectedContext({
+      selectedContext: SelectedContextType.Self,
+      orgs: [],
+    });
+
+    expect(result).toBe(true);
+  });
+
+  it('should return true when context is all', () => {
+    const result = userHasAccessToSelectedContext({
+      selectedContext: SelectedContextType.All,
+      orgs: [],
+    });
+
+    expect(result).toBe(true);
+  });
+
+  it('should return true when context id is present in orgs list', () => {
+    const result = userHasAccessToSelectedContext({
+      selectedContext: 1,
+      orgs: [
+        {
+          avatar_url: 'avatar_url',
+          description: '',
+          full_name: 'full_name',
+          id: 1,
+          location: '',
+          username: 'username',
+          website: '',
+        },
+      ],
+    });
+
+    expect(result).toBe(true);
+  });
+
+  it('should return false when context id is not present in orgs list', () => {
+    const result = userHasAccessToSelectedContext({
+      selectedContext: 2,
+      orgs: [
+        {
+          avatar_url: 'avatar_url',
+          description: '',
+          full_name: 'full_name',
+          id: 1,
+          location: '',
+          username: 'username',
+          website: '',
+        },
+      ],
+    });
+
+    expect(result).toBe(false);
   });
 });
